requirements.txt
fastapi
uvicorn
requests
python-multipart
openai-whisper
diffusers
transformers
accelerate
edge-tts
